refactor: Separate vibration examples into composable functions
- Created `singleVibrate.kt` for the single vibration example.
- Created `TimingVibrate.kt` for the timed/pattern vibration example.
- Updated `App.kt` to use these new composable functions and added a vertical scroll.
